:lipstick: better style for account page + hidden password in login page
