+++ Mobile/Carto/lib/views/widgets/form/form_fields/manager_information_form_field.dart
@@ -40,6 +40,8 @@ class _StringFormFieldState extends State<StringFormField> {
             ),
           ),
           errorText: _fieldError,
+          fillColor: Colors.white,
+          filled: true,
         ),
         autovalidateMode: AutovalidateMode.onUserInteraction,
         validator: _validator,
@@ -99,6 +101,7 @@ class _IntegerFormFieldState extends State<IntegerFormField> {
           ),
           errorText: _fieldError,
           fillColor: Colors.white,
+          filled: true,
         ),
         autovalidateMode: AutovalidateMode.onUserInteraction,
         validator: _validator,
