+++ Mobile/Carto/lib/views/widgets/form/other_fields/my_checkbox_list_tile.dart
@@ -39,13 +39,6 @@ class _MyCheckboxListTileState extends State<MyCheckboxListTile> {
       ),
       title: Text(widget.title),
       focusColor: Colors.white,
-
-        /*WidgetStateProperty.resolveWith((states) {
-          if (!states.contains(WidgetState.selected)) {
-            return Colors.red;
-          }
-          return null;
-        }),*/
     );
   }
 }
